Renamed files and rearanged packages
