+++ AndroidStudioCode/.idea/misc.xml
@@ -20,6 +20,17 @@
         <entry key="app/src/main/res/layout/layout_instructions.xml" value="0.1466438648995297" />
         <entry key="app/src/main/res/layout/layout_main.xml" value="0.1466438648995297" />
         <entry key="app/src/main/res/layout/layout_navegador.xml" value="0.3515625" />
+        <entry key="app/src/main/res/layouts/layout/layout_instructions.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_1_calculators/layout/layout_u1_p4_ohm.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_2_color_code_resistor/layout/layout_u2_p1_color.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_3_logic_gates/layout/layout_u3_p1_menu.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_3_logic_gates/layout/layout_u3_p2_logic_gates.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_3_logic_gates/layout/layout_u3_p3_truth_table.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_4_karnaugh/layout/layout_u4_p1_karnaugh.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_main/layout/layout_main.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_main/layout/layout_navigation_drawer.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_main/layout/layout_navigation_drawer_menu.xml" value="0.12141941000427534" />
+        <entry key="app/src/main/res/layouts/layouts_main/layout/layout_start.xml" value="0.12141941000427534" />
         <entry key="app/src/main/res/menu/draw_menu.xml" value="0.22" />
       </map>
     </option>
