+++ AndroidStudioCode/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="teamrocket.org.electronicadigital">
+    package="org.team_rocket_unc.electronica_digital_app">
 
     <application
         android:allowBackup="true"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ElectronicaDigital">
         <activity
-            android:name=".MainActivity"
+            android:name="org.team_rocket_unc.electronica_digital_app.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
