+++ AndroidStudioCode/app/src/main/java/org/team_rocket_unc/electronica_digital_app/units/unit_1_calculators/p3_led/LedCircuitToolFragment.java
@@ -1,0 +1,25 @@
+package org.team_rocket_unc.electronica_digital_app.units.unit_1_calculators.p3_led;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import org.team_rocket_unc.electronica_digital_app.R;
+
+public class LedCircuitToolFragment extends Fragment {
+
+    View view;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        view = inflater.inflate(R.layout.layout_u1_p3_led,container,false);
+        return view;
+    }
+
+}
