+++ AndroidStudioCode/app/src/main/java/org/team_rocket_unc/electronica_digital_app/units/unit_3_logic_functions/p1_logic_gates/LogicGatesToolFragment.java
@@ -1,0 +1,26 @@
+package org.team_rocket_unc.electronica_digital_app.units.unit_3_logic_functions.p1_logic_gates;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import org.team_rocket_unc.electronica_digital_app.R;
+
+public class LogicGatesToolFragment extends Fragment {
+
+    View view;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        view = inflater.inflate(R.layout.layout_u3_p2_logic_gates,container,false);
+        return view;
+    }
+
+}
+
