--- AndroidStudioCode/app/src/main/java/teamrocket/org/electronicadigital/H1Calculators/ledCircuitTool/LedCircuitToolFragment.java
@@ -1,25 +1,0 @@
-package teamrocket.org.electronicadigital.H1Calculators.ledCircuitTool;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import teamrocket.org.electronicadigital.R;
-
-public class LedCircuitToolFragment extends Fragment {
-
-    View view;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.layout_h1_led,container,false);
-        return view;
-    }
-
-}
