--- AndroidStudioCode/app/src/main/java/teamrocket/org/electronicadigital/H3LogicFunctions/logicGates/LogicGatesToolFragment.java
@@ -1,26 +1,0 @@
-package teamrocket.org.electronicadigital.H3LogicFunctions.logicGates;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import teamrocket.org.electronicadigital.R;
-
-public class LogicGatesToolFragment extends Fragment {
-
-    View view;
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.layout_h3_logic_gates,container,false);
-        return view;
-    }
-
-}
-
