+++ AndroidStudioCode/app/src/main/res/layouts/layouts_2_color_code_resistor/layout/layout_u2_p1_color.xml
