+++ AndroidStudioCode/app/src/main/res/layouts/layouts_3_logic_gates/layout/layout_u3_p2_logic_gates.xml
