+++ AndroidStudioCode/app/src/main/res/layouts/layouts_main/layout/layout_navigation_drawer_menu.xml
