+++ AndroidStudioCode/app/src/main/res/layouts/layouts_main/layout/layout_start.xml
